ClientAdvisor/App/requirements.txt
# Core requirements
azure-identity==1.20.0
# Flask[async]==2.3.2
openai==1.63.0
azure-search-documents==11.6.0b9
azure-storage-blob==12.24.1
python-dotenv==1.0.1
azure-cosmos==4.9.0
quart==0.20.0
uvicorn==0.34.0
gunicorn==23.0.0
aiohttp==3.11.12
quart-session==3.0.0
pymssql==2.3.2
httpx==0.28.1

# Linting and formatting tools
flake8==7.1.2
black==25.1.0
autoflake==2.3.1
isort==6.0.0

# Testing tools
pytest>=8.2,<9  # Compatible version for pytest-asyncio
pytest-asyncio==0.25.3
pytest-cov==6.0.0

pyodbc==5.2.0